+++ app/src/main/java/blockeq/com/stellarwallet/flowcontrollers/MnemonicFlowController.kt
@@ -1,0 +1,4 @@
+package blockeq.com.stellarwallet.flowcontrollers
+
+object MnemonicFlowController {
+}
