+++ app/src/main/res/layout/mnemonic_custon_view.xml
@@ -1,0 +1,16 @@
+<?xml version="1.0" encoding="utf-8"?>
+<ScrollView xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:app="http://schemas.android.com/apk/res-auto"
+    android:layout_width="match_parent"
+    android:layout_height="match_parent"
+    android:layout_below="@id/warningPhraseTextView">
+
+    <com.google.android.flexbox.FlexboxLayout
+        android:id="@+id/mnemonicGridView"
+        android:orientation="horizontal"
+        android:layout_width="match_parent"
+        android:layout_height="wrap_content"
+        android:layout_margin="@dimen/activity_vertical_margin"
+        app:flexWrap="wrap"/>
+
+</ScrollView>
